Modified more until works
With the section to ask if remember, the program now stores the marks of every single flashcard. Mark is plus one if user remember the answer while minus one if forget the answer. Also added a choice if going to edit marks as if the mark rise to 3 the review session will no longer show that flashcard anymore, by editing mark, enable user chance to review the flashcard that have been over 3marks before.